자동 주문 방향 문자열 수정: long과 short의 조건을 반전

+++ crypto-bot/strategy/hyper_disparity.py
@@ -279,7 +279,8 @@ def get_top5_consecutive_green(threshold=0.5):
 
                         # === 자동 주문 ===
                          
-                        direction_str = "long" if direction else "short"
+                        # direction_str = "long" if direction else "short"
+                        direction_str = "short" if direction else "long"
                         price = end_price
                         tp = price * (1.01 if direction_str == "long" else 0.99)
                         sl = price * (0.99 if direction_str == "long" else 1.01)
